+++ adapters/MoPub/MoPubAdapter/GADMediationAdapterMoPub.h
@@ -9,7 +9,11 @@ typedef NS_ENUM(NSInteger, GADMoPubErrorCode) {
   /// There was an error loading data from the network.
   GADMoPubErrorLoadingImages = 102,
   /// Missing server parameters.
-  GADMoPubErrorInvalidServerParameters = 103
+  GADMoPubErrorInvalidServerParameters = 103,
+  /// Banner size mismatch.
+  GADMoPubErrorBannerSizeMismatch = 104,
+  /// Banner ad size is smaller than minimum banner size set by the publisher.
+  GADMoPubErrorMinimumBannerSize = 105
 };
 
 @interface GADMediationAdapterMoPub : NSObject <GADMediationAdapter>